+++ Chimera/chimera.hpp
@@ -9,9 +9,9 @@ extern void ChimeraInit();
 
 namespace Chimera
 {
-  uint32_t millis();
-  void delayMilliseconds( uint32_t ms );
-  void delayMicroseconds( uint32_t us );
+  size_t millis();
+  void delayMilliseconds( size_t ms );
+  void delayMicroseconds( size_t us );
 }  // namespace Chimera
 
 #endif
